requirements.txt
fastapi[all]
Django
djangorestframework
djangorestframework-simplejwt
requests
django_filter
django-cors-headers
pyaudio
speechrecognition
python-dotenv
django-extensions
SQLAlchemy
python-dotenv